@@ -18,7 +18,7 @@
     "config": {
         "sort-packages": true,
         "platform": {
-            "php": "7.4.99"
+            "php": "8.1.99"
         },
         "allow-plugins": {
             "dealerdirect/phpcodesniffer-composer-installer": true,
@@ -29,17 +29,17 @@
     "extra": {
     },
     "require": {
-        "php": "^7.4 || ~8.0.0 || ~8.1.0 || ~8.2.0",
+        "php": "~8.1.0 || ~8.2.0 || ~8.3.0",
         "ext-ctype": "*",
         "ext-mbstring": "*"
     },
     "require-dev": {
-        "infection/infection": "^0.26.6",
-        "laminas/laminas-coding-standard": "~2.4.0",
+        "infection/infection": "^0.27.0",
+        "laminas/laminas-coding-standard": "~2.5.0",
         "maglnet/composer-require-checker": "^3.8.0",
-        "phpunit/phpunit": "^9.5.18",
-        "psalm/plugin-phpunit": "^0.17.0",
-        "vimeo/psalm": "^4.22.0"
+        "phpunit/phpunit": "^9.6.7",
+        "psalm/plugin-phpunit": "^0.18.4",
+        "vimeo/psalm": "^5.9"
     },
     "autoload": {
         "psr-4": {
