@@ -2,21 +2,18 @@
 
 namespace Drupal\viewsreference\Plugin\Field\FieldType;
 
-use Drupal\views\Views;
 use Drupal\Core\Entity\EntityTypeInterface;
 use Drupal\Core\Field\FieldStorageDefinitionInterface;
-use Drupal\Core\Field\PreconfiguredFieldUiOptionsInterface;
+use Drupal\Core\Field\Plugin\Field\FieldType\EntityReferenceItem;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\StringTranslation\TranslatableMarkup;
-use Drupal\Core\TypedData\OptionsProviderInterface;
-use Drupal\Core\Field\Plugin\Field\FieldType\EntityReferenceItem;
 use Drupal\Core\TypedData\DataDefinition;
+use Drupal\views\Views;
 
 /**
  * Defines the 'viewsreference' entity field type.
  *
- * Supported settings (below the definition's 'settings' key) are:
- * - target_type: The entity type to reference. Required.
+ * The target type for viewsreference fields should always be 'view'.
  *
  * @FieldType(
  *   id = "viewsreference",
@@ -25,12 +22,10 @@ use Drupal\Core\TypedData\DataDefinition;
  *   category = @Translation("Reference"),
  *   default_widget = "viewsreference_autocomplete",
  *   default_formatter = "viewsreference_formatter",
- *   list_class = "\Drupal\Core\Field\EntityReferenceFieldItemList",
+ *   list_class = "\Drupal\viewsreference\Plugin\Field\ViewsReferenceFieldItemList",
  * )
  */
-class ViewsReferenceItem extends EntityReferenceItem implements
-    OptionsProviderInterface,
-    PreconfiguredFieldUiOptionsInterface {
+class ViewsReferenceItem extends EntityReferenceItem {
 
   /**
    * {@inheritdoc}
@@ -48,6 +43,7 @@ class ViewsReferenceItem extends EntityReferenceItem implements
     return [
       'plugin_types' => ['block' => 'block'],
       'preselect_views' => [],
+      'enabled_settings' => [],
     ] + parent::defaultFieldSettings();
   }
 
@@ -56,23 +52,12 @@ class ViewsReferenceItem extends EntityReferenceItem implements
    */
   public static function propertyDefinitions(FieldStorageDefinitionInterface $field_definition) {
     $properties = parent::propertyDefinitions($field_definition);
-
     $properties['display_id'] = DataDefinition::create('string')
       ->setLabel(new TranslatableMarkup('Display Id'))
       ->setDescription(new TranslatableMarkup('The referenced display Id'));
-
-    $properties['argument'] = DataDefinition::create('string')
-      ->setLabel(new TranslatableMarkup('Argument'))
-      ->setDescription(new TranslatableMarkup('An optional argument or contextual filter to apply to the View'));
-
-    $properties['title'] = DataDefinition::create('string')
-      ->setLabel(new TranslatableMarkup('Title'))
-      ->setDescription(new TranslatableMarkup('Whether or not to include the View or Block title'));
-
     $properties['data'] = DataDefinition::create('string')
       ->setLabel(new TranslatableMarkup('Data'))
       ->setDescription(new TranslatableMarkup('Settings data for advanced use'));
-
     return $properties;
   }
 
@@ -90,27 +75,12 @@ class ViewsReferenceItem extends EntityReferenceItem implements
       // their IDs should not exceed the maximum length for bundles.
       'length' => $target_type_info->getBundleOf() ? EntityTypeInterface::BUNDLE_MAX_LENGTH : 255,
     ];
-
-    $schema['columns']['argument'] = [
-      'description' => 'An optional argument.',
-      'type' => 'varchar_ascii',
-      'length' => 255,
-    ];
-
-    $schema['columns']['title'] = [
-      'description' => 'Include title.',
-      'type' => 'int',
-      'length' => 11,
-    ];
-
     $schema['columns']['data'] = [
       'description' => 'Serialized data.',
       'type' => 'text',
       'size' => 'big',
     ];
-
     $schema['indexes']['display_id'] = ['display_id'];
-
     return $schema;
   }
 
@@ -120,26 +90,17 @@ class ViewsReferenceItem extends EntityReferenceItem implements
   public function setValue($values, $notify = TRUE) {
     // Select widget has extra layer of items.
     if (isset($values['target_id']) && is_array($values['target_id'])) {
-      $values['target_id'] = isset($values['target_id'][0]['target_id']) ? $values['target_id'][0]['target_id'] : NULL;
+      $values['target_id'] = $values['target_id'][0]['target_id'] ?? NULL;
     }
+
     // Empty string argument only possible if no argument supplied.
-    if (isset($values['argument']) && $values['argument'] === '') {
-      $values['argument'] = NULL;
+    $data = !empty($values['data']) ? unserialize($values['data'], ['allowed_classes' => FALSE]) : [];
+    if (isset($data['argument']) && '' === $data['argument']) {
+      $data['argument'] = NULL;
+      $values['data'] = serialize($data);
     }
+
     parent::setValue($values, FALSE);
-
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function isEmpty() {
-    $return = parent::isEmpty();
-    // Avoid loading the entity by first checking the 'display_id'.
-    if ($this->display_id === NULL || $this->display_id == '') {
-      return TRUE;
-    }
-    return $return;
   }
 
   /**
@@ -148,10 +109,12 @@ class ViewsReferenceItem extends EntityReferenceItem implements
   public function fieldSettingsForm(array $form, FormStateInterface $form_state) {
     $form = parent::fieldSettingsForm($form, $form_state);
     $settings = $this->getSettings();
-    $preselect_views = isset($settings['preselect_views']) ? $settings['preselect_views'] : [];
-    $default_plugins = isset($settings['plugin_types']) ? $settings['plugin_types'] : [];
-    $display_options = $this->getAllPluginList();
+    $preselect_views = $settings['preselect_views'] ?? [];
+    $default_plugins = $settings['plugin_types'] ?? [];
+    $enabled_settings = $settings['enabled_settings'] ?? [];
+    $display_options = $this->getAllViewDisplayIds();
     $view_list = $this->getAllViewsNames();
+
     $form['plugin_types'] = [
       '#type' => 'checkboxes',
       '#options' => $display_options,
@@ -162,27 +125,27 @@ class ViewsReferenceItem extends EntityReferenceItem implements
 
     $form['preselect_views'] = [
       '#type' => 'checkboxes',
-      '#title' => t('Preselect View Options'),
+      '#title' => $this->t('Preselect View Options'),
       '#options' => $view_list,
       '#default_value' => $preselect_views,
       '#weight' => 2,
     ];
 
-    return $form;
-  }
+    $enabled_settings_list = [];
+    $viewsreference_plugin_manager = \Drupal::service('plugin.manager.viewsreference.setting');
+    $plugin_definitions = $viewsreference_plugin_manager->getDefinitions();
+    foreach ($plugin_definitions as $plugin_definition) {
+      $enabled_settings_list[$plugin_definition['id']] = $plugin_definition['label'];
+    }
+    $form['enabled_settings'] = [
+      '#type' => 'checkboxes',
+      '#title' => $this->t('Enable extra settings'),
+      '#options' => $enabled_settings_list,
+      '#default_value' => $enabled_settings,
+      '#weight' => 3,
+    ];
 
-  /**
-   * Determines whether the item holds an unsaved entity.
-   *
-   * This is notably used for "autocreate" widgets, and more generally to
-   * support referencing freshly created entities (they will get saved
-   * automatically as the hosting entity gets saved).
-   *
-   * @return bool
-   *   TRUE if the item holds an unsaved entity.
-   */
-  public function hasNewEntity() {
-    return !$this->isEmpty() && $this->target_id === NULL && $this->entity->isNew();
+    return $form;
   }
 
   /**
@@ -190,17 +153,19 @@ class ViewsReferenceItem extends EntityReferenceItem implements
    */
   public static function getPreconfiguredOptions() {
     return [];
-
   }
 
   /**
-   * Helper function to get all display ids.
+   * Get all views display IDs.
+   *
+   * @return array
+   *   An array of view display IDs keyed bu plugin name.
    */
-  private function getAllPluginList() {
+  protected function getAllViewDisplayIds() {
     $types = Views::pluginList();
     $options = [];
     foreach ($types as $key => $type) {
-      if ($type['type'] == 'display') {
+      if ('display' === $type['type']) {
         $options[str_replace('display:', '', $key)] = $type['title']->render();
       }
     }
@@ -208,15 +173,29 @@ class ViewsReferenceItem extends EntityReferenceItem implements
   }
 
   /**
-   * Helper function to get all View Names.
+   * Get all enabled view names.
+   *
+   * @return array
+   *   An array of enabled view names keyed by view ID.
    */
-  private function getAllViewsNames() {
+  protected function getAllViewsNames() {
     $views = Views::getEnabledViews();
     $options = [];
     foreach ($views as $view) {
       $options[$view->get('id')] = $view->get('label');
     }
     return $options;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function isEmpty() {
+    // Avoid loading the entity by first checking the 'display_id'.
+    if (NULL === $this->display_id || '' == $this->display_id) {
+      return TRUE;
+    }
+    return parent::isEmpty();
   }
 
 }
