@@ -3,16 +3,16 @@
 namespace Drupal\viewsreference\Plugin\Field\FieldWidget;
 
 use Drupal\Core\Field\FieldItemListInterface;
+use Drupal\Core\Field\Plugin\Field\FieldWidget\EntityReferenceAutocompleteWidget;
 use Drupal\Core\Form\FormStateInterface;
 use Symfony\Component\Validator\ConstraintViolationInterface;
-use Drupal\Core\Field\Plugin\Field\FieldWidget\EntityReferenceAutocompleteWidget;
 
 /**
  * Plugin implementation of the 'entity_reference_autocomplete' widget.
  *
  * @FieldWidget(
  *   id = "viewsreference_autocomplete",
- *   label = @Translation("Views Reference Autocomplete"),
+ *   label = @Translation("Views reference autocomplete"),
  *   description = @Translation("An autocomplete views reference field."),
  *   field_types = {
  *     "viewsreference"
@@ -28,10 +28,7 @@ class ViewsReferenceWidget extends EntityReferenceAutocompleteWidget {
    */
   public function formElement(FieldItemListInterface $items, $delta, array $element, array &$form, FormStateInterface $form_state) {
     $element = parent::formElement($items, $delta, $element, $form, $form_state);
-
-    $test = ['filled' => TRUE];
-    $element = $this->fieldElement($element, $items, $delta, $test);
-
+    $element = $this->fieldElement($items, $delta, $element, $form, $form_state);
     return $element;
   }
 
@@ -39,17 +36,7 @@ class ViewsReferenceWidget extends EntityReferenceAutocompleteWidget {
    * {@inheritdoc}
    */
   public function errorElement(array $element, ConstraintViolationInterface $error, array $form, FormStateInterface $form_state) {
-    return isset($element['display_id']) ? $element['display_id'] : FALSE;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function massageFormValues(array $values, array $form, FormStateInterface $form_state) {
-    $values = parent::massageFormValues($values, $form, $form_state);
-    // Options creates an array which we need to flatten.
-    $values = $this->massageValues($values, $form, $form_state);
-    return $values;
+    return $element['display_id'] ?? FALSE;
   }
 
 }
