@@ -11,7 +11,7 @@ use Drupal\Core\Form\FormStateInterface;
  *
  * @FieldWidget(
  *   id = "viewsreference_select",
- *   label = @Translation("Views Reference Select list"),
+ *   label = @Translation("Views reference select list"),
  *   description = @Translation("An autocomplete views select list field."),
  *   field_types = {
  *     "viewsreference"
@@ -25,40 +25,20 @@ class ViewsReferenceSelectWidget extends OptionsSelectWidget {
   /**
    * {@inheritdoc}
    */
-  protected function getEmptyLabel() {}
-
-  /**
-   * {@inheritdoc}
-   */
   public function formElement(FieldItemListInterface $items, $delta, array $element, array &$form, FormStateInterface $form_state) {
-
     $select_element['target_id'] = parent::formElement($items, $delta, $element, $form, $form_state);
-
-    $select_element = $this->fieldElement($select_element, $items, $delta);
+    $select_element = $this->fieldElement($items, $delta, $select_element, $form, $form_state);
     $select_element['target_id']['#multiple'] = FALSE;
     if (!$this->isDefaultValueWidget($form_state)) {
       $selected_views = $items->getSetting('preselect_views');
-      $selected_views = array_diff($selected_views, ["0"]);
+      $selected_views = array_diff($selected_views, ['0']);
       $selected_views = $this->getViewNames($selected_views);
       if (count($selected_views) >= 1) {
-        $first_option = [$this->t("- None -")];
+        $first_option = ['_none' => $this->t('- Select a value -')];
         $select_element['target_id']['#options'] = array_merge($first_option, $selected_views);
       }
-      else {
-        $select_element['target_id']['#empty_option'] = $this->t('- None -');
-      }
     }
-
     return $select_element;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function massageFormValues(array $values, array $form, FormStateInterface $form_state) {
-    // Options creates an array which we need to flatten.
-    $values = $this->massageValues($values, $form, $form_state);
-    return $values;
   }
 
 }
